requirements.txt
requests
openai
selenium
webdriver-manager
python-dotenv